requirements.txt
beautifulsoup4==4.14.3
bs4==0.0.2
certifi==2025.11.12
charset-normalizer==3.4.4
idna==3.11
requests==2.32.5
soupsieve==2.8.1
typing_extensions==4.15.0
urllib3==2.6.2
neo4j
python-dotenv
langchain
langchain-community
langchain-classic
streamlit
langchain-neo4j
langchain_openai
langchain-huggingface
langchain-core
openai
sentence-transformers
FlagEmbedding
nltk
rank-bm25
playwright
ollama
huggingface_hub[hf_xet]
docling
ragas
spacy